Move dashboard and widget apps into api app
Consolidate dashboard, widget and api apps to
simplify code.
